+++ vtm/src/org/oscim/utils/geom/OBB2D.java
@@ -32,22 +32,22 @@ public class OBB2D {
             v[pos + 1] = y;
         }
 
-        public static float dot(float[] v, int a, int b) {
+        public static double dot(double[] v, int a, int b) {
             return v[a] * v[b] + v[a + 1] * v[b + 1];
         }
 
-        public final static float lengthSquared(float[] v, int pos) {
-            float x = v[pos + 0];
-            float y = v[pos + 1];
+        public final static double lengthSquared(double[] v, int pos) {
+            double x = v[pos + 0];
+            double y = v[pos + 1];
 
             return x * x + y * y;
         }
 
-        public final static void normalizeSquared(float[] v, int pos) {
-            float x = v[pos + 0];
-            float y = v[pos + 1];
+        public final static void normalizeSquared(double[] v, int pos) {
+            double x = v[pos + 0];
+            double y = v[pos + 1];
 
-            float length = x * x + y * y;
+            double length = x * x + y * y;
 
             v[pos + 0] = x / length;
             v[pos + 1] = y / length;
@@ -70,26 +70,26 @@ public class OBB2D {
             return (float) Math.sqrt(x * x + y * y);
         }
 
-        public final static void add(float[] v, int r, int a, int b) {
+        public final static void add(double[] v, int r, int a, int b) {
             v[r + 0] = v[a + 0] + v[b + 0];
             v[r + 1] = v[a + 1] + v[b + 1];
         }
 
-        public final static void sub(float[] v, int r, int a, int b) {
+        public final static void sub(double[] v, int r, int a, int b) {
             v[r + 0] = v[a + 0] - v[b + 0];
             v[r + 1] = v[a + 1] - v[b + 1];
         }
 
-        public final static void mul(float[] v, int pos, float a) {
+        public final static void mul(double[] v, int pos, float a) {
             v[pos + 0] *= a;
             v[pos + 1] *= a;
         }
     }
 
-    float originX;
-    float originY;
+    double originX;
+    double originY;
 
-    public final float[] vec = new float[4 * 2 + 2 * 2];
+    public final double[] vec = new double[4 * 2 + 2 * 2];
 
     // Corners of the box, where 0 is the lower left.
     //public final float[] corner = new float[ 4 * 2];
@@ -111,15 +111,15 @@ public class OBB2D {
     // Returns true if other overlaps one dimension of this.
     private boolean overlaps1Way(OBB2D other) {
         for (int a = 0; a <= 2; a += 2) {
-            float ax = vec[AXIS_X + a];
-            float ay = vec[AXIS_Y + a];
+            double ax = vec[AXIS_X + a];
+            double ay = vec[AXIS_Y + a];
 
             // dot product
-            float t = ax * other.vec[CORNER_X] + ay * other.vec[CORNER_Y];
+            double t = ax * other.vec[CORNER_X] + ay * other.vec[CORNER_Y];
 
             // Find the extent of box 2 on axis a
-            float tMin = t;
-            float tMax = t;
+            double tMin = t;
+            double tMax = t;
 
             for (int c = CORNER_X + 2; c < 8; c += 2) {
                 t = ax * other.vec[c] + ay * other.vec[c + 1];
@@ -191,13 +191,13 @@ public class OBB2D {
 
     }
 
-    public OBB2D(float cx, float cy, float width, float height, double acos, double asin) {
+    public OBB2D(double cx, double cy, double width, double height, double acos, double asin) {
 
-        float vx = (float) acos * width / 2;
-        float vy = (float) asin * width / 2;
+        double vx = (double) acos * width / 2;
+        double vy = (double) asin * width / 2;
 
-        float ux = (float) -asin * height / 2;
-        float uy = (float) acos * height / 2;
+        double ux = (double) -asin * height / 2;
+        double uy = (double) acos * height / 2;
 
         vec[CORNER_X] = cx + (vx - ux);
         vec[CORNER_Y] = cy + (vy - uy);
@@ -214,13 +214,13 @@ public class OBB2D {
         computeAxes();
     }
 
-    public void setNormalized(float cx, float cy, float vx, float vy, float width, float height,
-                              float dy) {
-        float ux = -vy;
-        float uy = vx;
+    public void setNormalized(double cx, double cy, double vx, double vy, double width, double height,
+                              double dy) {
+        double ux = -vy;
+        double uy = vx;
 
-        float hw = width / 2;
-        float hh = height / 2;
+        double hw = width / 2;
+        double hh = height / 2;
 
         if (dy != 0) {
             cx += vx * dy + vy * dy;
@@ -248,19 +248,19 @@ public class OBB2D {
         computeAxes();
     }
 
-    public void set(float cx, float cy, float dx, float dy, float width, float height) {
-        float vx = cx - dx;
-        float vy = cy - dy;
+    public void set(double cx, double cy, double dx, double dy, double width, double height) {
+        double vx = cx - dx;
+        double vy = cy - dy;
 
-        float a = (float) Math.sqrt(vx * vx + vy * vy);
+        double a = (double) Math.sqrt(vx * vx + vy * vy);
         vx /= a;
         vy /= a;
 
-        float hw = width / 2;
-        float hh = height / 2;
+        double hw = width / 2;
+        double hh = height / 2;
 
-        float ux = vy * hh;
-        float uy = -vx * hh;
+        double ux = vy * hh;
+        double uy = -vx * hh;
 
         vx *= hw;
         vy *= hw;
